refactor: extract useRecycleBin hook
Moves showEmptyBinConfirm and showMobileRecycleBin state plus the
undeleteTask, emptyRecycleBin, and confirmEmptyBin functions out of
App.jsx into a dedicated useRecycleBin hook. The recycleBin array
state remains at the App level (shared with useUndo and data
persistence) while the hook receives it as a parameter.
